copy file with io::copy instead of fs::copy fs::copy does not work with samba

+++ imgcopy/src/sync.rs
@@ -40,7 +40,7 @@ fn init_file_map(trg: &Path) -> Result<HashMap<[u8; 20], PathBuf>, ImgcpError> {
 fn prepare_path(mut target_path: PathBuf, file_name: &str) -> Result<PathBuf, ImgcpError> {
     if !target_path.exists() {
         fs::create_dir_all(&target_path)
-            .map_err(|source| ImgcpError::TargetDirNotCreated { source })?
+        .map_err(|source| ImgcpError::TargetDirNotCreated { source })?
     }
     target_path.push(file_name);
     let mut increment = 0;
@@ -52,15 +52,20 @@ fn prepare_path(mut target_path: PathBuf, file_name: &str) -> Result<PathBuf, Im
     Ok(target_path)
 }
 
-fn copy_file(src: &Path, trg: &Path, do_move: bool) -> Result<(), ImgcpError> {
-    fs::copy(src, trg)
-        .and_then(|r| {
-            if do_move {
-                fs::remove_file(src)?
+fn copy_file(src: &Path, trg: &Path, do_move: bool) -> Result<(), std::io::Error> {
+    let mut source = fs::File::open(src)?;
+    let mut target = fs::File::create(trg)?;
+    io::copy(&mut source, &mut target)
+    .and_then(|r| {
+        if do_move {
+            if r > 0 {
+            fs::remove_file(src)?
+            } else {
+                println!("{:?} was not moved! Is it empty?", src.to_path_buf());
             }
-            Ok(r)
-        })
-        .map_err(|source| ImgcpError::FileCopyFailed { source })?;
+        }
+        Ok(())
+    })?;
     Ok(())
 }
 
@@ -69,6 +74,7 @@ pub fn run(
     trg: &Path,
     do_move: bool,
     ignore_non_empty_target: bool,
+    verbose: bool,
 ) -> Result<(), ImgcpError> {
     let (src, trg) = checks::check_paths(src, trg, ignore_non_empty_target)?;
     let mut file_map = init_file_map(&trg)?;
@@ -129,8 +135,15 @@ pub fn run(
         };
 
         // finally move file into target folder
-        copy_file(entry.path(), &target_path, do_move)?;
-        println!("{:?} -> {:?}", entry.path(), &target_path);
+        if verbose {
+            println!("{:?} -> {:?}", entry.path(), &target_path);
+        }
+        copy_file(entry.path(), &target_path, do_move)
+          .map_err(|source| ImgcpError::FileCopyFailed {
+            source,
+            file: entry.into_path(),
+            trg: target_path.to_path_buf(),
+        })?;
         file_map.insert(hash, target_path.to_path_buf());
     }
     Ok(())
